backend/requirements.txt
# DSA Knowledge Base
qdrant-client>=1.7.0
openai>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Utilities
tavily-python>=0.7.14
redis>=5.0.0

# LangGraph / LangChain
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

